Change: attribute type for x, y, width, height from int to double

+++ src/Coin/Coin.cpp
@@ -1,6 +1,6 @@
 #include "Coin.hpp"
 
-Coin::Coin(int x, int y, int width, int height, int nilai) : Item(x, y, width, height, TYPE_COIN){
+Coin::Coin(double x, double y, double width, double height, int nilai) : Item(x, y, width, height, TYPE_COIN){
 	value = nilai;
 }
 
@@ -9,7 +9,7 @@ int Coin::getValue(){ //mendapatkan nilai koin
 } 
 
 void Coin::tick(int delay) {
-	int y = this->getY();
+	double y = this->getY();
 	if (!(this->isAtBottom)){
 		y = y + 1;
 		this->setY(y);
